Moves FormFieldUpload
