+++ src/components/Form/Col/Col.tsx
@@ -21,7 +21,7 @@ import FormFieldTable from "@root/components/Field/FormFieldTable";
 import FormFieldText from "@root/components/Field/FormFieldText";
 import FormFieldTextEditor from "@root/components/Field/FormFieldTextEditor";
 import FormFieldToggleSwitch from "@root/components/Field/FormFieldToggleSwitch";
-import FormFieldUpload from "@root/forms/FormFieldUpload";
+import FormFieldUpload from "@root/components/Field/FormFieldUpload";
 
 import evaluateShow from "@root/utils/show/evaluateShow";
 import RegisteredField from "../../Field/RegisteredField";
